backend/pyproject.toml
[project]
name = "physical-ai-textbook-backend"
version = "0.1.0"
description = "Backend API for Physical AI & Humanoid Robotics Textbook RAG chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "qdrant-client>=1.7.0",
    "openai>=1.3.5",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.7",
    "python-multipart>=0.0.6"
]